Updating EnchantmentManager for handling all spells

+++ Source/ACE.Server/Network/GameEvent/Events/GameEventMagicRemoveEnchantment.cs
@@ -2,7 +2,7 @@ namespace ACE.Server.Network.GameEvent.Events
 {
     public class GameEventMagicRemoveEnchantment : GameEventMessage
     {
-        public GameEventMagicRemoveEnchantment(Session session, uint spellID, uint layer)
+        public GameEventMagicRemoveEnchantment(Session session, ushort spellID, ushort layer)
             : base(GameEventType.MagicRemoveEnchantment, GameMessageGroup.UIQueue, session)
         {
             Writer.Write(spellID);
